Add files via upload
This python Script allows the analysis of cell migration data from simple coordinates and offers a nice visual with data and plots to showcase final results. It also allow the comparison of different given data. For more information reach out by private message if you plan to use this script for your analysis. Dependencies may have to be installed locally in a virtual environment for optimal performance.